move all tests to tests dir, move all classes to core dir
