+++ tests/test_board.py
@@ -1,0 +1,60 @@
+from unittest import TestCase, mock
+from unittest.mock import patch, MagicMock
+from core.board import Board
+
+
+class TestBoard(TestCase):
+    def test_board_is_formatted_correctly_with_empty_board(self):
+        board = Board()
+        initial_board = """
+    A  B  C  D  E  F  G  H  I  J
+ 1 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 2 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 3 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 4 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 5 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 6 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 7 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 8 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 9 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+10 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+"""
+        formatted_board = board.format()
+
+        self.assertEqual(formatted_board, initial_board)
+
+    def test_board_is_formatted_correctly_with_an_occupied_board(self):
+        board = Board()
+        board.update('A1')
+        occupied_board = """
+    A  B  C  D  E  F  G  H  I  J
+ 1 [X][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 2 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 3 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 4 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 5 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 6 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 7 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 8 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+ 9 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+10 [ ][ ][ ][ ][ ][ ][ ][ ][ ][ ]
+"""
+        formatted_board = board.format()
+
+        self.assertEqual(formatted_board, occupied_board)
+
+    def test_board_is_initialized_with_an_empty_state(self):
+       board = Board()
+       initial_board_state = [
+               [None, None, None, None, None, None, None, None, None, None],
+               [None, None, None, None, None, None, None, None, None, None],
+               [None, None, None, None, None, None, None, None, None, None],
+               [None, None, None, None, None, None, None, None, None, None],
+               [None, None, None, None, None, None, None, None, None, None],
+               [None, None, None, None, None, None, None, None, None, None],
+               [None, None, None, None, None, None, None, None, None, None],
+               [None, None, None, None, None, None, None, None, None, None],
+               [None, None, None, None, None, None, None, None, None, None],
+               [None, None, None, None, None, None, None, None, None, None],
+               ]
+       self.assertEqual(board.state, initial_board_state)
